Add support for helper methods in routes
Instead of helper blocks like in most Rack frameworks the helpers are defined inline in the routes. Since each route is a class that is a subclass of its parent route it also has access to all ancestors' helpers, but not siblings'.

This required class_exec instead of instance_exec.