Make it possible to abort the request handling from a before block
Calling Response#finish will stop the request handling and not call any further before blocks nor the handler. It will call all after blocks, though. The idea is that after blocks are for fixing up the response, for example encoding it, and that's still required even if the handling stopped early (for a concrete use case: a before block that sets the status to 404 and the body to an hash with an error key, you still want it to be serialized to JSON in an after block if that is how you serialize your responses).