Make sure before, after, rescue_from and middleware work when mounting
before, after and rescue_from blocks that are declared in the mounting app should be run for routes in the mounted app. Middleware used by the mounting app should also wrap routes in the mounted app.

The middleware is solved by overriding Route.new to wrap the app in middleware at creation time, while the before, after and rescue_froms are solved by appending them to the env hash while the request is being routed.