Rewrite the whole routing
This way the befores, afters, etc. don't need to be collected for each request, they're created at instantiation time. Middleware is also handled at instantiation time instead of the scary .new override thing.

The downside is that because of the middleware the request, response and app context need to be put into the env hash for the final call, but hopefully that can be gotten rid of for all non-middlewared routes.