Don't split parameters on semicolons, only on ampersands
Rack's default query parser considers `;` to be equivalent to `&` in query strings – which is how things worked in 1999, but not anymore.

Assuming that the [application/x-www-form-urlencoded spec][1] applies to GET requests in general `;` is not a separator anymore ([but it was in 1999][2]). This is at least what I've been able to gather from [this StackOverflow][3] post.

  [1]: https://www.w3.org/TR/2014/REC-html5-20141028/forms.html#url-encoded-form-data
  [2]: https://www.w3.org/TR/1999/REC-html401-19991224/appendix/notes.html#h-B.2.2
  [3]: https://stackoverflow.com/a/40768572/1109